add Guest to modal
