+++ src/App.js
@@ -6,7 +6,7 @@ import HeaderComponent from './components/HeaderComponent';
 import HelpComponent from './components/HelpComponent';
 import { Container } from 'react-bootstrap';
 import ProtectedRoute from './auth/protected-route';
-import ListPersonComponentCopy from './components/ListPersonComponentCopy';
+import ListPersonComponent from './components/ListPersonComponent';
 
 function App() {
   return (
@@ -16,7 +16,7 @@ function App() {
           <Switch>
             <Route exact path = "/" component = {HomeComponent}></Route>
             <Route path = "/login" component = {HomeComponent}></Route>
-            <ProtectedRoute path = "/students" component = {ListPersonComponentCopy}></ProtectedRoute>
+            <ProtectedRoute path = "/students" component = {ListPersonComponent}></ProtectedRoute>
             <Route path = "/help" component = {HelpComponent}></Route>
             <HomeComponent />
           </Switch>
