--- src/components/FooterComponent.jsx
@@ -1,15 +1,0 @@
-import React, { Component } from 'react';
-
-class FooterComponent extends Component {
-    render() {
-        return (
-            <div>
-                <footer className = "footer">
-                    <span className = "text-muted">All Rights Reserved 2022 @S3V</span>
-                </footer>
-            </div>
-        );
-    }
-}
-
-export default FooterComponent;
